+++ syntaxmp/src/commonMain/kotlin/com/gallatinapps/syntaxmp/compose/SyntaxRoleStyles.kt
@@ -1,4 +1,4 @@
-package com.gallatinapps.syntaxmp.compose.theme
+package com.gallatinapps.syntaxmp.compose
 
 import com.gallatinapps.syntaxmp.role.SyntaxRole